Created Favicon
The favicon animation runs one time and then stops as a word bubble that says "Chat" on it.  In some browsers, it may not animate at all, in which case it would remain as a blank word bubble.  This favicon was created with favicon.cc.  It is 16px x 16px, with 15 animation frames.