Fixed Bug
"Clear All" button was not available after it was clicked so that one person could not vote multiple times.  However, when the chat room was finally cleared, the button was not available to clear again.  Now this is fixed.  The button becomes visible again when the chat room is cleared.