Removed Theme-Changing Option
This option may be programmed in later, but it will not be available for now.  I do not currently know how to set CSS general styles with JavaScript, and there are more important things to focus on, such as removing the Kihtrak Cloud Variable Library and Using actual websockets for data transfer.