Finished Administration Page
This page allows any administrators of the Chat website to control the data - that is, they may clear the data when it gets out of hand, send an important message or notice to everyone on the website, view all the Chat data, and change the password that allows them to do all these things.  People who enter the wrong password will be blocked from the admin page until the `exception` is enabled by the programming host.