Added Message Broadcaster
This function enables the administrators to, from the admin.html page, send everybody on the website a message.  This message will appear in an alert dialog on everybody's screen.  A good example may be: "The server will be closed for maintenance in two minutes.  After about three minutes of maintenance, it should be running again".