Linked Proper Functions to Button Clicks
Some functions were left undefined after some removing of code from Chat version 1.0.  `proposeNewRoom()` has now been changed to the new function that does the same thing, which is `page('new')`.  `newRoom()` has now been replaced with its proper function as well, which is the `room.create()` function.  (The two functions are different because the first goes to the page where you create a new chat room and the second is the button in that page to confirm the entered room name and password and officially make the chat room and open it.